requirements.txt
sentence-transformers
transformers
datasets
faiss-cpu
streamlit
ragas
torch
scikit-learn
pandas
numpy
